Misc
Filtering of axes is now optional. Relocated charCount and
anchorCharCount maps so that they are local to just the
applyTextLengthBias function and not global to the transform function.
Added some more axes. Refactored findBestElement to avoid repeatedly
querying the score of the current best element when iterating over all
elements. Improved some comments