Feed now shows up in subs list after subscribing
Fixed a bug where the feed object was not properly passed as a property
of the message that was sent upon subscribing. This is only a partial
fix as there is an issue with how a new feed is added regarding tracking
new entries and passing these to the callback. For now we just fix this
in the options page to suppress the errors.