Moved image helper functions related to polling
Moved functions from image-utils.js into poll.js, this was the only
calling context. Even though the functions decrease the coherency of
poll.js, it is the only place these functions are called and they are
highly specific and not generally useful anywhere else so they belong
in poll.js (I think).