Fixed a bug with setImageDimensions
It did not consider the case where the document did not contain any
images, so it would leave the poll process in a pending state. Now this
checks for 0 images and exits properly, which leads to polling
completing as expected. I only found this bug because this was the
first time in testing I encountered a page without images. It turns out
that nytimes.com appends images with JS, which I don’t evaluate, so no
script nytimes.com articles appear without any images.