Fixed image fetch bug in poll.js
fetchImageDimensions now properly waits to callback until images that
need to be fetched have been fetched. Previously it was calling back
multiple times simply when an internal for loop completed. This is
easily avoided by not incremented the processed image count until after
fetched images have been fetched. Also, I changed the condition from >=
to == because it will never be >. Also, I moved isObjectURL to utils
and renamed it.