Refactored poll.js to use db.updateFeed instead of db.putFeed
Part of a series of revisions focusing on having db.js focus only on db
calls and no added functionality, and for having more atomic functions
with singular purposes. Added poll.createdMergedFeed because it is no
longer the responsibility of db.updateFeed or db.putFeed. I also
renamed pollContext to context because I don’t think it is ambiguous,
and inlined poll.onFetchFeed into poll.start. I also removed the check
for whether feed.urls is defined for a locally stored feed, because it
is now safe to assume all feeds have a urls property containing at
least one url.