Properly cascade favicon info from feed to entries when polling
After the recent changes to use the new favicon service, I no longer
send feed link and let the view find the favicon from the link. Instead
I use the new FaviconService to get the favicon based on the feed’s
link, and now I cascade this property to the entry.