Fixed minor bug in FaviconService.prototype.findIconURLInDocument
This now checks if href value is truthy. If I don’t, then the new URL
constructor gets an empty href and defaults to returning the base URL,
which is incorrect. Instead this now only tries to create the url if
the href value is defined and not an empty string. This way the service
does not mistakenly return a non-favicon-url url as the result url.