Improvements to polling
Added entry and feed helper classes. Simplified dependency linking in
call to polling in alarms.js. Improved logging in favicon-service and
fixed a minor bug with an undefined variable and a previously never
evaluated branch. feed-http-service now also yields the response
document so that it can be used for other processing, like favicon
processing.