Fixed a bug with setting image dimensions
I was incrementing numProcessed in the wrong location. I moved it here
in a prior commit recently, turns out it was wrong. It should only be
incremented after image is fetched, not before, because otherwise the
callback condition is met more than once, leading to multiple
callbacks, which leads to strange behavior in polling.