Improvements to Subscription.add and FeedCache.addFeed
Simplified the callback of Subscription.add. It is now the options page
responsibility to determine the content of the message to show.
OPMLImportService no longer passes a callback to addFeed. The callback
parameter to addFeed is now optional. addFeed now sets the id of the
feed and returns the added feed, instead of requiring the caller to
find it from the event property.