Added additional debugging to searchGoogleFeedsOnResponse
I was testing subscribe and got an error when searching, so I added
some additional debugging to try and hunt it down. I think it was just
a net hiccup, but it lead to a null response, and I wasn’t satisfied
with the amount of information provided to debug it. Using dir will
list all the available information so the next time I observe the error
I can maybe find out something.