Fixed a variable access error in subscribeFindFeed
After adding the check for whether a feed with a similar URL exists in
a prior commit, and adding the connection parameter to subscribe, I
moved some code into the subscribeFindFeed function so that it could
have multiple entry points, either via the provided connection or the
optional connection. However, after moving, I forgot to update how to
access the connection variable, because now it is only available via
the context, and not as a function argument.