Partial reversion to functions
I only need DI for things that need to be changed, so I think the task
approach was wrong. It never felt right in the first place. Passing in
dependencies as function parameters is better, and I only need to pass
in dependencies that need a different state in testing and production.