[favicon] Check reachability of favicons
Issue #284. The lookup function now also also checks that any icon urls
found when searching the input document can also be successfully
fetched. Previously this just returned whatever icon url was found
without verifying whether there exists a retrievable resource at the
given url. This change should reduce the number of errors that occur in
the UI when displaying a favicon, where the app thinks it has a valid
favicon but in reality it does not.