[subscribe] fix subscribe errors
Add missing await call when setting favicon, otherwise this sets a
promise as a feed property which then causes update-feed to fail. In
addition, always set id in update-feed because for some reason sometimes
it fails to set the id of the feed as expected when creating a new feed.
In addition, implement a basic subscribe test.