[create-feed-test] renamed write-new-feed-test
So, initially I thought I was just testing a particular use of
write-feed, but there is no need to avoid a bit of abstraction. If
write-feed is what is used to create a feed, then it is ok to name the
test differently. I want to name the test after the particular use, not
just the function call variant.