[model-access] trap channel errors
Calling postMessage can throw an InvalidStateError when the channel is closed. The channel calls all occur in a deferred context from within a Promise executor. Such errors are uncaught and leave the promise in an unsettled state. Such errors are not automatically translated into rejections. Therefore, use try/catch. However, instead of explicitly rejecting, log errors as a side effect, reserving rejection for database errors. It would otherwise be misleading to throw an error when in fact the database was modified. This means that all model access methods do not provide a message broadcast guarantee.