[model-access] drop useless parameters from iterateEntries
The mode parameter was no longer important once archiveEntries was implemented because archiveEntries now does its own internal iteration, leaving only one mode for iterateEntries to ever consider. The writable parameter was not useful because in 100% of use cases it was true. Basically writable was an invariant, and invariants generally should not be function parameters.