[sanitize-document] decouple from fetch policy module
Instead of directly importing the module and hardcoding the policy, the policy function is now a parameter as a component of the options parameter. This is part of a series of changes for moving sanitize-document to general filters and out of app folder. I think this is the last change before move.