[color-contrast-filter] create folder
Promote color-contrast-filter from a helper module to sandoc into its own global module. Merge css-color into the color-contrast-filter folder.

Note that I moved the test as well, because I want to start merging in tests, but this violates some other conventions so there is still more work to be done regarding the test.