[coerce-element] move into its own module
The issue with dom.js is that most of the functions in, while conceptually related to the dom, share fundamentally different purposes. dom.js is just completely the wrong approach. Furthermore, only one other module uses coerce-element, and it does not make use of the rest of dom, so this solves that bundling issue and now we get the banana without the jungle.