Move is-alphanumeric from string.js to url-utils.js
url-utils was the only user of this function. moving this function removed the coupling of url-utils to string, and increased the coherency of string, because is-alphanumeric was unrelated to the other functions and did not call or get called by them, the only reason it was there was because it similarly operated on a string argument. functional coherency is better than logical coherency.