[export-opml] shift download concern to UI
export-opml sits in a lower layer and should not have UI-related code. A download is a UI action. So basically I split this module up into two layers, and this now represents only the lower layer. The higher layer code now sits in the one place it was called in left-panel.js