[db] add missing import
The subscribe test failed saying db.validate_feed was not a function. I inspected db.js and lo and behold, sanity is not re-exported. In the transition to using a sinle api module in the form of db.js I forgot to expose sanity.js functions and somehow managed to not encounter this error until fixing the subscribe test.