[base] deprecate
Partial progress on migration from layered architecture to modular 
architecture. It took some time for me to grasp the concept but the two 
are really not the same thing. Layering is implicit in a modular 
architecture, there is no need to enforce explicit boundaries such as by 
using folders representing layers.