[dom-filters] created image-reachable-filter
Added a new filter that tries to fetch images in a document and removes any that it could not fetch. First draft. Enhanced the image size filter to account for the already-performed fetch and attempt to reuse the fetch result by checking some special element attributes. Added a test for the new filter. Revised some of the test utilities to account for some annoying requirements in set-base-uri, which at the moment, forbids the use of chrome-extension:// urls.

See #725.

Test: ran the new test a few times