[set-base-uri] simplifications
Deprecate has-valid-base-uri. It was wrong to test against chrome-extension protocol, and this was the only value add of this function over a simple test against the document.baseURI property. This left document-utils with only set-base-uri left. Rename the module to set-base-uri because the extra layer of abstraction was no longer beneficial.