[db] remove OO from types
Redesign feed/entry entities. Do away with the object-oriented approach. 
OO introduced a need for deserialization where otherwise there was no 
need. That is a lot of perf drop just for some syntactic sugar. Maybe in 
a language like C++ it is worth it using some kind of template, but in 
JS it is all just a waste. So the decision to use indexedDB has some 
pretty dramatic consequences.