[db][delete-feed] stop using feed index
Revise delete-feed so that iteration over the entries store looking for 
entries belonging to the feed no longer makes use of the feed index on 
the entry store. delete-feed was the sole user of the feed index. Using 
the index did speed up the operation, and it did avoid loading full 
entry data into memory. However, there is no need to be fast here. This 
is a rarely used operation where there is no expectation of speed. 
Because this is the sole client of the feed index, it should now be 
possible in a subsequent commit to simply remove the feed index, which 
should improve the performance of other operations that need good 
performance.

I also took this opportunity to break up the function a bit so that it 
is more readable.

Progress on #651.

Test: delete-feed test passes