[db][entry] be explicit about archived-date property
archive-entries introduces a property used nowhere else, and I missed it 
when trying to make all properties explicit in the entry module. Also, I 
renamed the field from date-archived to archived-date to be consistent. 
Note this means that older entries have a bad field. For now this is ok 
because nothing uses this field.

Tests: none