[db][archive-entries] improve testing
Actually flesh out the test. Create some archivable entries, run 
archive-entries, and verify the state of the database after the 
operation.

In addition, I encountered the increased-size warning when implementing 
the test. I modified the archive-entry helper in archive-entries to also 
remove empty properties. The problem was that once I began explicitly 
listing properties in the constructor, and using the constructor, I 
forgot that now these keys exist in the object, and as a result, 
archive-entry was creating many empty properties.

This mostly completes #582.

Tests: ran the updated archive-entries test, it passes