[db] deprecate set-entry-read-state
After implementing patch-entry and modifying set-entry-read-state to 
delegate to patch-entry, set-entry-read-state's value addition was so 
marginalized that I decided to just have the caller call patch-entry 
directly and bypass it entirely. It was always an overly-specific use 
case anyway, so I think it was the right decision.

Progress on #800

Tests: slideshow reading articles works as expected