Switch from appendChild to append
Use the new method introduced by HTML5. There are a few cases where I am 
still not taking full advantage of it (e.g. still using createTextNode 
or .textContent=) but I will eventually get to those.

This completes #808.

Tests: ran a few of the tests and played around in the view for a bit.