[resource-utils] fix resource type validation bug
A place in the UI marks articles as read, which eventually calls into 
patch-resource with a set of delta properties, which eventually passes 
that props object into resource-utils.validate. validate expects type to 
be either feed or entry, but type may not be in the delta set. For now, 
allow type to be undefined, so that validation passes. this makes 
marking articles as read work again.