[subscribe] properly set resource type
In the refactor away from feed objects and to using resource objects, I 
introduced a new type property. However, I forgot to include a step in 
the subscribe function when it creates a basic feed object, it needs to 
set the type to feed, because it is no longer a feed but a resource. Now 
the type is properly set.

Tests: subscribe test now passes