[subscribe] clarify the type of the object created
subscribe internally used outdated nomenclature. by declaring the 
variable with the name resource, it is more obvious that the object 
being created is a resource and not a feed. Had I done this originally I 
probably would have avoided the mistake that I had to correct in the 
antecedent commit where because the object represents a resource it must 
have a type field, which is not a requirement shared by the now 
antiquated feed object.

Tests: subscribe test still passes