[archive-resources] move to services
In a past commit I modified archive-resources to use multiple 
transactions and therefore not require direct access to raw indexedDB 
transactions and requests in order to avoid the indexedDB transaction 
timeout issue. This makes the implementation basically just call into 
some other db functions. While this is a module that is primarily 
concerned with data, it is also viewable as a higher level service. It 
feels more natural as a service.

Test: ran via cli and observed that it worked without error.